ckc_web/requirements.txt
asgiref==3.2.7
Django==3.0.5
django-modeltranslation==0.15
numpy==1.18.3
Pillow==7.1.2
pytz==2020.1
six==1.14.0
sqlparse==0.3.1
